dreamos-agent-trainer/requirements.txt
# Dream.OS Agent Trainer Requirements
# Core dependencies for the training pipeline

# Data processing and ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# Text processing and embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
tokenizers>=0.13.0

# Configuration and utilities
pyyaml>=6.0
pathlib2>=2.3.7

# Database and storage
sqlite3  # Built-in with Python

# Optional: For training (requires GPU)
# torch>=2.0.0
# datasets>=2.10.0
# peft>=0.4.0
# accelerate>=0.20.0

# Optional: For advanced features
# faiss-cpu>=1.7.0  # For vector similarity search
# chromadb>=0.4.0   # Alternative vector database
# openai>=1.0.0     # For OpenAI embeddings